Move login check into services
