+++ src/js/engagement_view/src/components/Dashboard.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from "react";
 import Button from "@material-ui/core/Button";
 import { useAsync } from "react-async-hook";
-import { checkLogin } from "../Login";
+import { checkLogin } from "../services/loginService";
 import { Link } from "react-router-dom";
 import { dasboardStyles } from "./makeStyles/DashboardStyles";
 import { LoginNotification, GraplHeader } from "./reusableComponents";
