+++ src/js/engagement_view/src/services/loginService.ts
@@ -1,6 +1,7 @@
 import apiFetch from "./fetch";
 import DEV_API_EDGES from "./constants";
 
-export const checkLogin = (): Promise<boolean | null> => {
-  return apiFetch(`${DEV_API_EDGES.auth}/checkLogin`, "post");
+export const checkLogin = async (): Promise<boolean | null> => {
+  const loginData= await apiFetch(`${DEV_API_EDGES.auth}/checkLogin`, "post");
+  return loginData['success'] === 'True';
 };
